+++ gems/aws-sdk-frauddetector/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.30.0 (2022-01-26)
+------------------
+
+* Feature - Added new APIs for viewing past predictions and obtaining prediction metadata including prediction explanations: ListEventPredictions and GetEventPredictionMetadata
+
 1.29.0 (2021-12-21)
 ------------------