+++ gems/aws-sdk-sagemaker/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.116.0 (2022-01-26)
+------------------
+
+* Feature - API changes relating to Fail steps in model building pipeline and add PipelineExecutionFailureReason in PipelineExecutionSummary.
+
 1.115.0 (2022-01-05)
 ------------------